+++ parlai/core/worlds.py
@@ -44,7 +44,6 @@ All worlds are initialized with the following parameters:
 import copy
 import random
 import time
-import importlib
 
 from functools import lru_cache
 from typing import List, Dict, Any
@@ -58,7 +57,7 @@ from parlai.core.agents import (
     create_agents_from_shared,
     create_task_agent_from_taskname,
 )
-from parlai.core.loader import load_task_module
+from parlai.core.loader import load_task_module, load_world_module
 from parlai.core.metrics import aggregate_metrics
 from parlai.utils.misc import Timer, display_messages
 from parlai.core.opt import Opt
@@ -750,58 +749,6 @@ def _override_opts_in_shared(table, overrides):
     return table
 
 
-# WORLD LOADER
-def load_world_module(
-    taskname: str,
-    interactive_task: bool,
-    num_agents: int,
-    default_world=None
-):
-    sp = taskname.strip()
-    repo = 'parlai'
-    if sp.startswith('internal:'):
-        # To switch to local repo, useful for non-public projects
-        # (make a directory called 'parlai_internal' with your private agents)
-        repo = 'parlai_internal'
-        sp = sp[9:]
-    sp = sp.split(':')
-    if '.' in sp[0]:
-        # The case of opt['task'] = 'parlai.tasks.squad.agents:DefaultTeacher'
-        # (i.e. specifying your own path directly, assumes DialogPartnerWorld)
-        if default_world is not None:
-            world_class = default_world
-        elif num_agents == 2:
-            world_class = DialogPartnerWorld
-        else:
-            world_class = MultiAgentDialogWorld
-    else:
-        task = sp[0].lower()
-        if len(sp) > 1:
-            sp[1] = sp[1][0].upper() + sp[1][1:]
-            world_name = sp[1] + "World"
-            if interactive_task:
-                world_name = "Interactive" + world_name
-        else:
-            if interactive_task:
-                world_name = "InteractiveWorld"
-            else:
-                world_name = "DefaultWorld"
-        module_name = "%s.tasks.%s.worlds" % (repo, task)
-        try:
-            my_module = importlib.import_module(module_name)
-            world_class = getattr(my_module, world_name)
-        except (ModuleNotFoundError, AttributeError):
-            # Defaults to this if you did not specify a world for your task.
-            if default_world is not None:
-                world_class = default_world
-            elif num_agents == 2:
-                world_class = DialogPartnerWorld
-            else:
-                world_class = MultiAgentDialogWorld
-
-    return world_class
-
-
 class BatchWorld(World):
     """
     BatchWorld contains many copies of the same world.
@@ -1324,52 +1271,6 @@ def _create_task_agents(opt: Opt):
     return task_agents
 
 
-# def _get_task_agents(opt: Opt, user_agents, default_world=None):
-#     task_agents = _create_task_agents(opt)
-#     sp = opt['task'].strip()
-#     repo = 'parlai'
-#     if sp.startswith('internal:'):
-#         # To switch to local repo, useful for non-public projects
-#         # (make a directory called 'parlai_internal' with your private agents)
-#         repo = 'parlai_internal'
-#         sp = sp[9:]
-#     sp = sp.split(':')
-#     if '.' in sp[0]:
-#         # The case of opt['task'] = 'parlai.tasks.squad.agents:DefaultTeacher'
-#         # (i.e. specifying your own path directly, assumes DialogPartnerWorld)
-#         if default_world is not None:
-#             world_class = default_world
-#         elif len(task_agents + user_agents) == 2:
-#             world_class = DialogPartnerWorld
-#         else:
-#             world_class = MultiAgentDialogWorld
-#     else:
-#         task = sp[0].lower()
-#         if len(sp) > 1:
-#             sp[1] = sp[1][0].upper() + sp[1][1:]
-#             world_name = sp[1] + "World"
-#             if opt.get('interactive_task', False):
-#                 world_name = "Interactive" + world_name
-#         else:
-#             if opt.get('interactive_task', False):
-#                 world_name = "InteractiveWorld"
-#             else:
-#                 world_name = "DefaultWorld"
-#         module_name = "%s.tasks.%s.worlds" % (repo, task)
-#         try:
-#             my_module = importlib.import_module(module_name)
-#             world_class = getattr(my_module, world_name)
-#         except (ModuleNotFoundError, AttributeError):
-#             # Defaults to this if you did not specify a world for your task.
-#             if default_world is not None:
-#                 world_class = default_world
-#             elif len(task_agents + user_agents) == 2:
-#                 world_class = DialogPartnerWorld
-#             else:
-#                 world_class = MultiAgentDialogWorld
-#     return task_agents
-
-
 def create_task_world(opt: Opt, user_agents, default_world=None):
     """
     Instantiate a world with the supplied options and user agents.
@@ -1379,7 +1280,7 @@ def create_task_world(opt: Opt, user_agents, default_world=None):
     task_agents = _create_task_agents(opt)
     world_class = load_world_module(
         opt['task'],
-        opt['interactive_task'],
+        opt.get('interactive_task', False),
         len(user_agents + task_agents),
         default_world=default_world
     )
